requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
enum34==1.1.10
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
namedlist==1.8
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.3
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-xdist==3.5.0
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
